error handler

+++ app/Http/Controllers/MockResourceController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use \Illuminate\Database\QueryException;
 use App\Models\Mock;
 use App\Http\Resources\MockResource;
 
@@ -10,23 +11,31 @@ class MockResourceController extends Controller
 {
     public function get($endpoint){
         $requestedMock = Mock::where('endpoint', $endpoint)->first();
-        return new MockResource($requestedMock);
+        return (!is_null($requestedMock)) ? new MockResource($requestedMock) : abort(404);
     }
 
     public function post($endpoint, Request $request){
         $mockedRequest = new Mock;
 
         $mockedRequest->endpoint = $endpoint;
-        $mockedRequest->query = json_encode($request-> query());
-        $mockedRequest->payload = json_encode($request-> post());
+        $mockedRequest->query = json_encode($request->query());
+        $mockedRequest->payload = json_encode($request->post());
         
 
-        if($mockedRequest->save())
-            return new MockResource($mockedRequest);
+        try{
+            if($mockedRequest->save())
+                return new MockResource($mockedRequest);
+        } catch(QueryException $e) {
+            abort(403);
+        }
+        
     }
 
     public function put($endpoint, Request $request){
         $requestedMock = Mock::where('endpoint', $endpoint)->first();
+
+        if(is_null($requestedMock))
+            return abort(404);
 
         $requestedMock->query = json_encode($request->query());
         $requestedMock->payload = json_encode($request->post());
@@ -38,6 +47,9 @@ class MockResourceController extends Controller
     public function patch($endpoint){
         $requestedMock = Mock::withTrashed()->where('endpoint', $endpoint)->first();
 
+        if(is_null($requestedMock))
+            return abort(404);
+
         $requestedMock->deleted_at = null;
         
         if($requestedMock->save())
@@ -47,6 +59,9 @@ class MockResourceController extends Controller
     public function delete($endpoint, Request $request){
         $requestQuery = $request->query();
         $requestedMock = Mock::where('endpoint', $endpoint)->first();
+
+        if(is_null($requestedMock))
+            return abort(404);
         
         if(array_key_exists('MockAPiForceDelete', $requestQuery) && $requestQuery['MockAPiForceDelete'] == 'true'){
             $requestedMock->forceDelete();
